+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [0.26.2] - 2026-01-10
+### Added
+- `ray_cluster_logging_config.enabled` optional subfield. Setting is tied to `var.addons_config.ray_operator_enabled`
+- `ray_cluster_monitoring_config.enabled` optional subfield, but is currently set to `false` as this requires Google Cloud Managed Service for Prometheus
+- `examples/kuberay`
+
 ## [0.26.1] - 2026-01-07
 ### Changed
 - Updated Taskfiles